Initial commit
Initial commit of The Dude.

This is pretty advanced for a first commit. It is not actually the first
commit but as the project started out as a hack-about, in the early
stages of development things were committed to the repo that should not
have been such as api keys and passwords.

As The Dude has developed these things have gradually been removed
however they were still available in the history so for safety's sake,
the original git repository has been archived and a new one created so
that the project can be pushed to github.