Added TheDude.command method
Added a method to find a command by its original expression. This was
needed because in the commands hash they are stored by processed regex
to make them easier to iterate and match over.

This required some refactoring of expressions in order to store both the
original and processed expression.